refactor(a2ui): reorganize catalog renderer imports and separate basic catalog renderer
